Simplified how UI flags are stored and read. * Implemented '%' values for position attribute. ** The actual value calculated uses the total unused width/height in the parent container.

+++ source/code/systems/ui_system/public/ice/ui_types.hxx
@@ -18,7 +18,7 @@ namespace ice::ui
     struct Action;
 
     enum class ResourceType : ice::u32;
-    enum class ElementFlags : ice::u16;
+    enum class ElementFlags : ice::u32;
     enum class ElementType : ice::u8;
